@@ -33,78 +33,80 @@
     "@expo/react-native-action-sheet": "^3.13.0",
     "@procivis/react-native-components": "git+https://gitlab+deploy-token-1:ZDtywH2vsxxN4_xheZyY@gitlab.procivis.ch/procivis/components/react-native.git#5ed09744d6c50cdcf0309b6b2faa0df88895c939",
     "@react-native-async-storage/async-storage": "^1.14.1",
-    "@react-native-community/image-editor": "^2.3.0",
-    "@react-native-community/masked-view": "0.1.10",
     "@react-native/normalize-color": "^2.0.0",
-    "@react-navigation/bottom-tabs": "^6.0.9",
-    "@react-navigation/native": "~6.0.1",
-    "@react-navigation/native-stack": "^6.0.2",
-    "@sentry/react-native": "^4.1.2",
+    "@react-navigation/bottom-tabs": "^6.5.7",
+    "@react-navigation/native": "~6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
+    "@sentry/react-native": "^5.4.1",
     "apisauce": "2.0.0",
     "buffer": "^6.0.3",
-    "expo": ">=44.0.0-0 <45.0.0",
-    "expo-local-authentication": "^12.1.1",
-    "expo-localization": "^12.0.1",
     "i18n-js": "3.8.0",
-    "lottie-react-native": "^4.1.3",
+    "lottie-react-native": "^5.1.5",
     "mobx": "6.1.8",
     "mobx-react-lite": "3.2.0",
     "mobx-state-tree": "5.0.1",
-    "react": "17.0.2",
-    "react-native": "0.67.5",
+    "react": "18.2.0",
+    "react-native": "0.71.7",
     "react-native-blur-overlay": "^1.0.7",
-    "react-native-bootsplash": "^3.2.7",
+    "react-native-bootsplash": "^4.6.0",
     "react-native-camera": "^4.2.1",
     "react-native-device-info": "^10.0.2",
     "react-native-draggable-flatlist": "^3.1.0",
     "react-native-fs": "^2.18.0",
-    "react-native-gesture-handler": "1.10.3",
+    "react-native-gesture-handler": "2.9.0",
     "react-native-get-random-values": "^1.7.0",
+    "react-native-keep-awake": "^4.0.0",
     "react-native-keychain": "^8.0.0",
+    "react-native-localize": "^3.0.0",
     "react-native-picker": "git+https://gitlab+deploy-token-9:F2skvy6yGB5XhEf1oivN@gitlab.procivis.ch/procivis/components/react-native-picker.git#8a834ce43a70622b53995ed968342a494c8cbfb9",
-    "react-native-reanimated": "^2.9.0",
-    "react-native-safe-area-context": "3.1.8",
-    "react-native-screens": "^3.11.0",
+    "react-native-reanimated": "^2.17.0",
+    "react-native-safe-area-context": "4.5.2",
+    "react-native-screens": "^3.20.0",
     "react-native-svg": "^12.1.1",
-    "react-native-ultimate-config": "^3.4.2",
+    "react-native-touch-id": "^4.4.1",
+    "react-native-ultimate-config": "^6.0.0",
     "react-native-url-polyfill": "^1.3.0",
     "timeago.js": "^4.0.2",
     "validate.js": "0.13.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@babel/plugin-proposal-decorators": "7.12.1",
-    "@babel/plugin-proposal-optional-catch-binding": "7.12.1",
-    "@babel/runtime": "^7.12.5",
+    "@babel/core": "^7.21.0",
+    "@babel/preset-env": "^7.20.0",
+    "@babel/runtime": "^7.21.0",
+    "@react-native-community/eslint-config": "^3.2.0",
+    "@tsconfig/react-native": "^2.0.2",
     "@types/i18n-js": "3.0.3",
-    "@types/jest": "26.0.19",
-    "@types/react": "17.0.39",
-    "@types/react-native": "0.66.15",
-    "@types/react-test-renderer": "17.0.1",
-    "@typescript-eslint/eslint-plugin": "5.30.0",
-    "@typescript-eslint/parser": "5.30.0",
-    "babel-jest": "26.6.3",
+    "@types/jest": "^29.2.1",
+    "@types/react": "^17.0.41",
+    "@types/react-native": "0.71.6",
+    "@types/react-native-keep-awake": "^2.0.3",
+    "@types/react-native-touch-id": "^4.4.0",
+    "@types/react-test-renderer": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "5.59.2",
+    "@typescript-eslint/parser": "5.59.2",
+    "babel-jest": "^29.2.1",
     "babel-loader": "8.2.2",
-    "eslint": "7.15.0",
-    "eslint-config-prettier": "7.0.0",
-    "eslint-config-standard": "16.0.2",
-    "eslint-plugin-import": "2.22.1",
+    "deprecated-react-native-prop-types": "^4.1.0",
+    "eslint": "^8.19.0",
+    "eslint-config-prettier": "8.8.0",
+    "eslint-plugin-import": "^2.27.0",
+    "eslint-plugin-jest": "^27.2.1",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-react": "7.21.5",
+    "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-native": "3.10.0",
+    "eslint-plugin-react-native": "^4.0.0",
     "eslint-plugin-simple-import-sort": "^8.0.0",
     "fbjs-scripts": "3.0.0",
     "husky": "^7.0.4",
-    "jest": "^25.5.4",
-    "jest-expo": "^40.0.1",
+    "jest": "^29.2.1",
     "jetifier": "1.6.6",
     "license-checker-rseidelsohn": "^3.1.0",
+    "metro-react-native-babel-preset": "^0.73.7",
     "npm-run-all": "4.1.5",
     "patch-package": "6.4.7",
     "postinstall-prepare": "1.0.1",
-    "prettier": "2.2.1",
+    "prettier": "^2.4.1",
     "react-devtools-core": "4.10.1",
     "react-dom": "^17.0.2",
     "react-native-clean-project": "^3.6.3",
@@ -114,19 +116,20 @@
     "shelljs": "^0.8.4",
     "solidarity": "2.3.1",
     "sort-json": "^2.0.0",
-    "typescript": "4.7.4"
+    "typescript": "4.8.4"
   },
   "jest": {
-    "preset": "jest-expo",
+    "preset": "react-native",
     "setupFiles": [
-      "<rootDir>/test/setup.ts"
+      "<rootDir>/test/setup.ts",
+      "<rootDir>/node_modules/react-native-gesture-handler/jestSetup.js"
     ],
     "testPathIgnorePatterns": [
       "/node_modules/",
       "/e2e"
     ],
     "transformIgnorePatterns": [
-      "node_modules/(?!(jest-)?react-native|@react-native|@react-native-async-storage|@react-navigation|@storybook|@react-native-community|rn-material-ui-textfiel|expo-)"
+      "node_modules/(?!(jest-)?react-native|@react-native|@react-native-async-storage|@react-navigation|@react-native-community|rn-material-ui-textfiel|react-native-screens|react-native-reanimated|expo-)"
     ],
     "moduleNameMapper": {
       "\\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/test/mock-file.ts",
