@@ -1,2 +1,2 @@
-module.exports = 'SvgMock';
-module.exports.ReactComponent = 'SvgMock';
+export default 'SvgMock';
+export const ReactComponent = 'SvgMock';
