@@ -3,4 +3,4 @@
  *
  * @param ms The number of milliseconds to wait.
  */
-export const delay = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
+export const delay = (ms: number) => new Promise<void>((resolve) => setTimeout(resolve, ms));
