@@ -10,11 +10,13 @@ jest.doMock('react-native', () => {
       Image: {
         ...ReactNative.Image,
         resolveAssetSource: jest.fn((source) => mockFile), // eslint-disable-line @typescript-eslint/no-unused-vars
-        getSize: jest.fn((
-          uri: string, // eslint-disable-line @typescript-eslint/no-unused-vars
-          success: (width: number, height: number) => void,
-          failure?: (error: unknown) => void, // eslint-disable-line @typescript-eslint/no-unused-vars
-        ) => success(100, 100)),
+        getSize: jest.fn(
+          (
+            uri: string,
+            success: (width: number, height: number) => void,
+            failure?: (error: unknown) => void, // eslint-disable-line @typescript-eslint/no-unused-vars
+          ) => success(100, 100),
+        ),
       },
     },
     ReactNative,
